#  Cache results of model capability detection