This PR implements caching feature requested in https://github.com/taketwo/llm-ollama/pull/16#issuecomment-2795108365.

Without this PR:

```bash
hyperfine --warmup 5 "llm models"
Benchmark 1: llm models
  Time (mean ± σ):     843.0 ms ±   5.9 ms    [User: 543.9 ms, System: 77.6 ms]
  Range (min … max):   835.6 ms … 851.1 ms    10 runs
```

With this PR:

```bash
hyperfine --warmup 5 "llm models"
Benchmark 1: llm models
  Time (mean ± σ):     618.1 ms ±   5.5 ms    [User: 539.7 ms, System: 76.2 ms]
  Range (min … max):   609.4 ms … 630.4 ms    10 runs
```

My Ollama instance has 6 models, so saving approximately 40 ms per model.

@iamjackg please give this a try and let me know!